moved DeckFromTsv to configuration, because is was dumb to have to configuration packages. Renamed to DefaultDeck, because we no longer use the Tsv file
